# api/swagger/openapi.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "title": "NATS Horizon API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "version": "1.0"
  },
  "basePath": "/api",
  "paths": {
...
    "/streams": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "List all streams",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/StreamResponse"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "Create a new stream",
        "parameters": [
          {
            "description": "Stream creation request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateStreamRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/StreamResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/streams/{name}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "Get stream details",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StreamResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "Update a stream",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "description": "Stream update request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStreamRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StreamResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "Delete a stream",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SuccessResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
    "/streams/{name}/purge": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "streams"
        ],
        "summary": "Purge messages from a stream",
        "parameters": [
          {
            "type": "string",
            "description": "Stream name",
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "description": "Purge options",
            "name": "request",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/PurgeStreamRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PurgeStreamResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateStreamRequest": {
      "type": "object",
      "required": [
        "name",
        "storage",
        "subjects"
      ],
      "properties": {
        "max_age": {
          "type": "string"
        },
        "max_bytes": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "replicas": {
          "type": "integer",
          "maximum": 5,
          "minimum": 1
        },
        "retention": {
          "type": "string",
          "enum": [
            "limits",
            "interest",
            "workqueue"
          ]
        },
        "storage": {
          "type": "string",
          "enum": [
            "file",
            "memory"
          ]
        },
        "subjects": {
          "type": "array",
          "minItems": 1,
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    },
...
    "PurgeStreamRequest": {
      "type": "object",
      "properties": {
        "sequence": {
          "type": "integer"
        },
        "subject": {
          "type": "string"
        }
      }
    },
    "PurgeStreamResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "remaining": {
          "type": "integer"
        }
      }
    },
...
    "StreamConfigResponse": {
      "type": "object",
      "properties": {
        "max_age": {
          "type": "string"
        },
        "max_bytes": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "replicas": {
          "type": "integer"
        },
        "retention": {
          "type": "string"
        },
        "storage": {
          "type": "string"
        },
        "subjects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "StreamResponse": {
      "type": "object",
      "properties": {
        "config": {
          "$ref": "#/definitions/StreamConfigResponse"
        },
        "state": {
          "$ref": "#/definitions/StreamStateResponse"
        }
      }
    },
    "StreamStateResponse": {
      "type": "object",
      "properties": {
        "bytes": {
          "type": "integer"
        },
        "consumers": {
          "type": "integer"
        },
        "delete_marks": {
          "type": "integer"
        },
        "first_seq": {
          "type": "integer"
        },
        "first_ts": {
          "type": "string"
        },
        "last_seq": {
          "type": "integer"
        },
        "last_ts": {
          "type": "string"
        },
        "messages": {
          "type": "integer"
        },
        "num_pending": {
          "type": "integer"
        }
      }
    },
...
    "SuccessResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "UpdateStreamRequest": {
      "type": "object",
      "properties": {
        "max_age": {
          "type": "string"
        },
        "max_bytes": {
          "type": "integer"
        },
        "replicas": {
          "type": "integer",
          "maximum": 5,
          "minimum": 1
        },
        "subjects": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}